Update BT
Move BT activity to DeviceView
