--- app/src/main/res/layout/bluetooth.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:id="@+id/BTSettings"
-    android:weightSum="1">
-
-    <ListView
-        android:layout_width="match_parent"
-        android:layout_height="281dp"
-        android:id="@+id/BTDeviceList"
-        style="@style/NoTitle"
-        android:clickable="true"
-        android:visibility="visible"
-        android:headerDividersEnabled="false"
-        android:choiceMode="singleChoice"
-        android:background="#ffacc3c4"
-        android:layout_weight="0.11" />
-
-</LinearLayout>
