+++ app/src/main/res/layout/device.xml
@@ -3,13 +3,15 @@
     android:orientation="vertical" android:layout_width="match_parent"
     android:layout_height="match_parent" android:background="@color/white">
 
-    <TextView
-        android:text="设备"
+    <ListView
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id="@+id/test3"
-        android:textColor="#000"
-        android:layout_marginRight="3dp"
-        android:gravity="center_vertical|center"
-        android:textSize="8pt"/>
+        android:layout_height="281dp"
+        android:id="@+id/BTDeviceList"
+        style="@style/NoTitle"
+        android:clickable="true"
+        android:visibility="visible"
+        android:headerDividersEnabled="false"
+        android:choiceMode="singleChoice"
+        android:background="#ffacc3c4"
+        android:layout_weight="0.11" />
 </LinearLayout>
